⚡️ Change default data type of SOCPA

+++ demo_second_order.py
@@ -31,6 +31,6 @@ target_plain = np.array([bytearray.fromhex(p) for p in plain_text], dtype=np.uin
 estimated_power = power_modeling_for_cpa_16_using_cuda(target_plain, None, 0, 4, 16, True, benchmark=True)
 print("* Power modeling finished.")
 
-cor_mat = cpa_cuda_16_bit(target_trace, estimated_power, benchmark=True, dtypes=('double', 'float', 'double'))
+cor_mat = cpa_cuda_16_bit(target_trace, estimated_power, benchmark=True, dtypes=('double', 'double', 'float'))
 print(f"* Guessed key is 0x{np.argmax(np.max(abs(cor_mat), axis=1)):04X}. "
       f"({np.max(np.max(abs(cor_mat), axis=1)):.04f})")
